Project restructuring: modular structure in core/, vpn/, and udp/, adjusted imports, added __init__.py files, and updated requirements.txt
