+++ assets/utilizadores.txt
@@ -1,0 +1,1 @@
+admin;admin123;admin
